backend/requirements.txt
fastapi
pydantic
uvicorn

youtube-transcript-api
groq
python-dotenv
fastapi-jwt-auth
pydantic==1.10.13
